refactor: rename profile styles and update import paths for modals
